Storeのモジュール分割
