+++ client/src/router/index.js
@@ -4,7 +4,7 @@ import Login from '@/pages/Login.vue';
 import Home from '@/pages/Home.vue';
 import User from '@/pages/User.vue';
 import UserHome from '@/components/UserHome.vue';
-import UserList from '@/components/UserList.vue'
+import UserList from '@/components/UserList.vue';
 import store from '../store';
 
 Vue.use(VueRouter);
@@ -40,7 +40,7 @@ const routes = [
         path: 'all',
         name: 'userList',
         component: UserList,
-      }
+      },
     ],
   },
 ];
